Linux 3.6 compat: Remove `struct zone` hack
Linux and Solaris both define `struct zone`, but for different purposes.
In Linux, `struct zone` is used to describe different regions of memory
while on Solaris, `struct zone` is used to describe containers. This
caused a conflict that was originally worked around by renaming the
Linux definitions to `struct Xzone`. This caused a build failure on
Linux 3.6. The Solaris code declares `struct zone` as a convention, but
in reality, it depends on the typedef zone_t, not `struct zone`. As a
consequence, we can workaround this issue by making `struct zone` into a
nameless struct, which permits the build process to succeed on Linux
3.6.

Signed-off-by: Richard Yao <ryao@gentoo.org>