Use Linux infrastructure for dtrace_xcall()
Comments suggest that the rationale for using custom code for
dtrace_xcall() was that it on_each_cpu did not interact well with the
custom locking. Since that code needs to be replaced with more
conventional Linux locks, the rationale for a custom xcall
implementation is gone. Furthermore, it is not clear that this interacts
well with on_each_cpu(), so lets drop it in favor of wrapping
on_each_cpu(). This avoids having to worry about strange interactions.

Signed-off-by: Richard Yao <ryao@gentoo.org>