Use Linux interrupt management functions
The use of custom code makes debugging difficult and there does not
appear any reason for it given that the native Linux functions work on
ARM, so lets get rid of it.

This resolves potential issues where Linux's memory allocation functions
fail to warn that we are in an interrupt context. Memory allocations in
such contexts are unsafe and do not happen on any other DTrace platform,
so it is a very good thing to ensure that we are warned when this
happens.

Signed-off-by: Richard Yao <ryao@gentoo.org>